Enable shared code between /web and /server
Until now, we've been able to share types from /common to both our
/server and /web codebases. However, importing any code into /server
from /common that wasn't stripped out by TypeScript (so, anything but
types), caused our entire directory structure to change. It broke the
build in a non-obvious way, and has been a frustrating limitation.

I've slowly and surely been upgrading our Webpack / TypeScript
dependencies, which was part of the puzzle.

**This is where specific info will go, once it's done. For now it's just notes.**

_These are rough notes I've kept for myself while working on the project. I've changed directions several times, so not all of these sentences are accurate. I've tried to delete anything that's no-longer true at all, but there are some half-truths here._

The big change here is that we've moved from only relying on [Yarn workspaces](https://classic.yarnpkg.com/en/docs/workspaces/) on the server to also including [TypeScript path mapping](https://www.typescriptlang.org/docs/handbook/module-resolution.html#path-mapping).

Yarn workspaces expect us to build /common as its own module, and then import it. Right now, we're using /common as a collection of loose files. This works with Node, but only if the built files end up in the project's root (so, no compile folder like `dist` or `js`).

Whether or not we use the symlink functionality of Yarn workspaces, our root `package.json` will still have a `workspaces` key. That's because it lets us consolidate our `yarn.lock`, and run `yarn install` just once.

In the process of doing this work I realized that we have a lot of options for solving this task. Here are a few off the top of my head:

- Yarn Workspaces
- Yarn / NPM Private Packages
- Yarn link
- "file: ../" dependencies
- TypeScript Project References
- TypeScript Internal Modules
- TypeScript Path Aliases
- Lerna
- Oh… have I not even mentioned webpack yet?

We already using some of these, which made the task frustrating. The goal is to pick a subset that works for us, and strictly stick to that subset.

- Fixme: server output is happening twice somehow. Outputting the raw files, then also server/ and common/ folders.
- Fixme: Remove all `@ts-ignore` lines.
- Todo: can we bring this back? "I also prefixed the `common` alias with `@`, to make it clear that it's an alias. I've updated all import statements to use `@common/*`".
- Todo: ensure bundle sizes haven't exploded.
- Todo: Ensure nothing has gotten in the way of our IDE-ish TypeScript capabilities.
- Future: https://nodejs.org/api/esm.html#esm_package_exports comes with Node 13. Stable versions are even-numbered, which should be out summer 2020. Will simplify things if we choose to keep common/ files un-bundled.

**/end notes**

This commit also adds some trivial `real code` imports into /server,
so we can test it out right away.

## Test plan:

For quick checks, I found the following adequate:

`cd server && tsc -b`

or,

`cd common && tsc`

To run a final test of this change, I wanted to be sure that things would work
from scratch. The test plan below is a bit overkill, and it will take a while,
but it was worth the peace of mind.

```
# Scorch the earth!
docker ps -a
# …`docker kill` / `docker rm` all the common voice containers listed above.
docker images
# …`docker rmi` all the common voice images listed above. Woah.
rm -rf web/node_modules/ server/node_modules/ node_modules/ web/dist/ server/js common/js

# Now we truly start from a clean slate.
# Note to self: flaky builds here, sometimes expects a local `yarn install`?
docker-compose up
```

Did it work? Hope so!

## Relevant history:

The current monorepo structure was initially introduced in
https://github.com/mozilla/voice-web/commit/edcf9ae56948b517ce505c6af7a975f0790b3db7